+++ src/coreclr/src/CMakeLists.txt
@@ -1,5 +1,4 @@
 include_directories("inc")
-include_directories("strongname/inc")
 include_directories("inc/winrt")
 include_directories("debug/inc")
 include_directories("debug/inc/${ARCH_SOURCES_DIR}")
@@ -67,7 +66,6 @@ add_subdirectory(vm)
 add_subdirectory(md)
 add_subdirectory(debug)
 add_subdirectory(inc)
-add_subdirectory(strongname)
 add_subdirectory(binder)
 add_subdirectory(classlibnative)
 add_subdirectory(dlls)
