+++ src/coreclr/src/gc/sample/gcenv.ee.cpp
@@ -96,11 +96,7 @@ uint32_t CLREventStatic::Wait(uint32_t dwMilliseconds, bool bAlertable)
     return result;
 }
 
-#ifndef __GNUC__
-__declspec(thread) Thread * pCurrentThread;
-#else // !__GNUC__
 thread_local Thread * pCurrentThread;
-#endif // !__GNUC__
 
 Thread * GetThread()
 {
