+++ src/coreclr/src/ilasm/CMakeLists.txt
@@ -77,7 +77,6 @@ set(ILASM_LINK_LIBRARIES
   mdcompiler_wks
   mdruntime_wks
   mdruntimerw_wks
-  strongname_tool
   mdstaticapi
   ${END_LIBRARY_GROUP} # End group of libraries that have circular references
   ceefgen
