+++ src/coreclr/src/ildasm/exe/CMakeLists.txt
@@ -70,7 +70,6 @@ set(ILDASM_LINK_LIBRARIES
     mdcompiler_wks
     mdruntime_wks
     mdruntimerw_wks
-    strongname_tool
     mdstaticapi
     ${END_LIBRARY_GROUP} # End group of libraries that have circular references
     corguids
