+++ src/coreclr/src/inc/predeftlsslot.h
@@ -19,7 +19,7 @@
 // them.
 enum PredefinedTlsSlots
 {
-    TlsIdx_StrongName,
+    TlsIdx_OwnedCrstsChain, // slot to store the Crsts owned by this thread
     TlsIdx_JitPerf,
     TlsIdx_JitX86Perf,
     TlsIdx_JitLogEnv,
@@ -33,7 +33,6 @@ enum PredefinedTlsSlots
 
     // Add more indices here.
     TlsIdx_ThreadType, // bit flags to indicate special thread's type
-    TlsIdx_OwnedCrstsChain, // slot to store the Crsts owned by this thread
     TlsIdx_CantAllocCount, //Can't allocate memory on heap in this thread
 
     // A transient thread value that indicates this thread is currently walking its stack
