+++ src/coreclr/src/inc/strongnameinternal.h
@@ -8,7 +8,13 @@
 #ifndef _STRONGNAME_INTERNAL_H
 #define _STRONGNAME_INTERNAL_H
 
-#include <strongname.h>
+// Public key blob binary format.
+typedef struct {
+    unsigned int SigAlgID;       // (ALG_ID) signature algorithm used to create the signature
+    unsigned int HashAlgID;      // (ALG_ID) hash algorithm used to create the signature
+    ULONG        cbPublicKey;    // length of the key in bytes
+    BYTE         PublicKey[1];   // variable length byte array containing the key value in format output by CryptoAPI
+} PublicKeyBlob;
 
 // Determine the number of bytes in a public key
 DWORD StrongNameSizeOfPublicKey(const PublicKeyBlob &keyPublicKey);
@@ -19,5 +25,12 @@ bool StrongNameIsValidPublicKey(const PublicKeyBlob &keyPublicKey);
 // Determine if a public key is the ECMA key
 bool StrongNameIsEcmaKey(__in_ecount(cbKey) const BYTE *pbKey, DWORD cbKey);
 bool StrongNameIsEcmaKey(const PublicKeyBlob &keyPublicKey);
+
+HRESULT StrongNameTokenFromPublicKey(BYTE* pbPublicKeyBlob,  // [in] public key blob
+    ULONG    cbPublicKeyBlob,
+    BYTE** ppbStrongNameToken,     // [out] strong name token
+    ULONG* pcbStrongNameToken);
+
+VOID StrongNameFreeBuffer(BYTE* pbMemory);
 
 #endif // !_STRONGNAME_INTERNAL_H
