--- src/coreclr/src/strongname/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_subdirectory(api)
