+++ src/coreclr/src/vm/gdbjit.cpp
@@ -15,11 +15,7 @@
 #include "gdbjit.h"
 #include "gdbjithelpers.h"
 
-#ifndef __GNUC__
-__declspec(thread) bool tls_isSymReaderInProgress = false;
-#else // !__GNUC__
 thread_local bool tls_isSymReaderInProgress = false;
-#endif // !__GNUC__
 
 #ifdef _DEBUG
 static void DumpElf(const char* methodName, const char *addr, size_t size)
